Fixed the inconsistent size issue on resize and the type animation malfunction on the home screen.
Once a 3D object is created but it is not part of the current screen view and the screen is resized,
the object not part of the current screen does not get resized. And so, it remains the same when it is
reintroduced to the current view screen. We fix it by resizing these objects before reintroducing them
to the current view screen.

On the other hand, the type animation malfunctions when the animation is stopped during the cursor blinking
stage and it is restarted before the cursor blinking stage ends and move on to the backspace stage. This
results in typewriting and backspacing happening at the same moment in time causing the glitch. While this
is a source code issue and we will have to patch it or let the owners of the package take care of the issue.
For now, we fix it by reducing the cursor blinking stage's time duration called 'backDelay' in order to
prevent the above mentioned issue from ever occuring in a humanly manner.